+++ evrmore/core/key.py
@@ -627,6 +627,6 @@ class CPubKey(bytes):
             return '%s(b%s)' % (self.__class__.__name__, super(CPubKey, self).__repr__())
 
 __all__ = (
-        'CECKey',
-        'CPubKey',
+    'CECKey',
+    'CPubKey',
 )
